requirements.txt
asgiref==3.3.1
Django==3.1.5
django-appconf==1.0.4
django-compressor==2.4
django-sass-processor==0.8.2
django-solo==1.1.5
libsass==0.20.1
Pillow==8.1.0
pytz==2020.5
rcssmin==1.0.6
rjsmin==1.1.0
six==1.15.0
sqlparse==0.4.1
